format fix

+++ BSB_LAN/BSB_LAN.ino
@@ -3858,7 +3858,7 @@ char *build_pvalstr(bool extended) {
     if(roundf(decodedTelegram.prognr * 10) != roundf(decodedTelegram.prognr) * 10)
       len+=sprintf_P(outBuf, PSTR("%4.1f "), decodedTelegram.prognr);
     else
-      len+=sprintf_P(outBuf, PSTR("%d "), roundf(decodedTelegram.prognr));
+      len+=sprintf_P(outBuf, PSTR("%d "), (int)roundf(decodedTelegram.prognr));
 
     len+=strlen(strcpy_PF(outBuf + len, decodedTelegram.catdescaddr));
     len+=strlen(strcpy_P(outBuf + len, PSTR(" - ")));
